requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Djoser==2.2.2
psycopg2-binary==2.9.9
pillow==10.1.0
django-cors-headers==4.3.1
celery[redis]==5.2.7
flower==1.2.0
boto3==1.33.9
pybase64==1.3.1
django-storages==1.14.2